fix(kmip-wasm): gate the wasm32 certificate-projection call site correctly
register_import_export.rs's Register/Import call into
certify::project_certificate_to_engine unconditionally, but the certify
module is #[cfg(feature = "native")]-gated (its CSR/rcgen issuance logic
doesn't cross-compile to wasm32). No one had attempted a wasm32 build
of the kmip crate since certificate support was added, so this was
silently broken.

Extracts the wasm-safe projection helper (no rcgen/aws_lc_rs dependency)
into its own always-compiled cert_projection module; certify.rs re-exports
it so its own native call sites (bootstrap_ca_certificate, store_certificate)
are unaffected. Verified via native cargo test (full kmip suite green) and
a successful wasm32 build + in-wasm smoke test.
